git-svn-id: https://svn.alfresco.com/repos/alfresco-enterprise/alfresco/HEAD/root@30956 c4b6b30b-aa2e-2d43-bbcb-ca4b014f7261

+++ source/java/org/alfresco/repo/content/caching/cleanup/CachedContentCleaner.java
@@ -19,14 +19,20 @@
 package org.alfresco.repo.content.caching.cleanup;
 
 import java.io.File;
+import java.util.Date;
+import java.util.concurrent.locks.ReentrantReadWriteLock;
 
 import org.alfresco.repo.content.caching.CacheFileProps;
 import org.alfresco.repo.content.caching.ContentCacheImpl;
 import org.alfresco.repo.content.caching.FileHandler;
+import org.alfresco.repo.content.caching.quota.UsageTracker;
 import org.alfresco.util.Deleter;
+import org.apache.commons.io.FileUtils;
 import org.apache.commons.logging.Log;
 import org.apache.commons.logging.LogFactory;
 import org.springframework.beans.factory.annotation.Required;
+import org.springframework.context.ApplicationEventPublisher;
+import org.springframework.context.ApplicationEventPublisherAware;
 
 /**
  * Cleans up redundant cache files from the cached content file store. Once references to cache files are
@@ -34,17 +40,113 @@ import org.springframework.beans.factory.annotation.Required;
  * 
  * @author Matt Ward
  */
-public class CachedContentCleaner implements FileHandler
+public class CachedContentCleaner implements FileHandler, ApplicationEventPublisherAware
 {
     private static final Log log = LogFactory.getLog(CachedContentCleaner.class);
     private ContentCacheImpl cache;   // impl specific functionality required
+    private long minFileAgeMillis = 0;
     private Integer maxDeleteWatchCount = 1;
-    
-    public void execute()
-    {   
-        cache.processFiles(this);
+    private ReentrantReadWriteLock lock = new ReentrantReadWriteLock();
+    private boolean running;
+    private UsageTracker usageTracker;
+    private long newDiskUsage;
+    private long numFilesSeen;
+    private long numFilesDeleted;
+    private long sizeFilesDeleted;
+    private long numFilesMarked;
+    private Date timeStarted;
+    private Date timeFinished;
+    private ApplicationEventPublisher eventPublisher;
+    private long targetReductionBytes;
+   
+    /**
+     * This method should be called after the cleaner has been fully constructed
+     * to notify interested parties that the cleaner exists.
+     */
+    public void init()
+    {
+        eventPublisher.publishEvent(new CachedContentCleanerCreatedEvent(this));
     }
     
+    public void execute()
+    {
+        execute("none specified");
+    }
+    
+    public void executeAggressive(String reason, long targetReductionBytes)
+    {
+        this.targetReductionBytes = targetReductionBytes;
+        execute(reason);
+        this.targetReductionBytes = 0;
+    }
+    
+    public void execute(String reason)
+    {
+        lock.readLock().lock();
+        try
+        {
+            if (running)
+            {
+                // Do nothing - we only want one cleaner running at a time.
+                return;
+            }
+        }
+        finally
+        {
+            lock.readLock().unlock();
+        }
+        lock.writeLock().lock();
+        try
+        {
+            if (!running)
+            {
+                if (log.isInfoEnabled())
+                {
+                    log.info("Starting cleaner, reason: " + reason);
+                }
+                running = true;
+                resetStats();
+                timeStarted = new Date();
+                cache.processFiles(this);
+                timeFinished = new Date(); 
+                
+                if (usageTracker != null)
+                {
+                    usageTracker.setCurrentUsageBytes(newDiskUsage);
+                }
+                
+                running = false;
+                if (log.isInfoEnabled())
+                {
+                    log.info("Finished, duration: " + getDurationSeconds() + "s, seen: " + numFilesSeen +
+                                ", marked: " + numFilesMarked +
+                                ", deleted: " + numFilesDeleted +
+                                " (" + String.format("%.2f", getSizeFilesDeletedMB()) + "MB, " +
+                                sizeFilesDeleted + " bytes)" +
+                                ", target: " + targetReductionBytes + " bytes");
+                }
+            }
+        }
+        finally
+        {
+            lock.writeLock().unlock();
+        }
+    }
+    
+    
+    /**
+     * 
+     */
+    private void resetStats()
+    {
+        newDiskUsage = 0;
+        numFilesSeen = 0;
+        numFilesDeleted = 0;
+        sizeFilesDeleted = 0;
+        numFilesMarked = 0;
+    }
+
+
     @Override
     public void handle(File cachedContentFile)
     {
@@ -52,36 +154,83 @@ public class CachedContentCleaner implements FileHandler
         {
             log.debug("handle file: " + cachedContentFile);
         }
+        numFilesSeen++;
+        CacheFileProps props = null;
+        boolean deleted = false;
         
-        CacheFileProps props = null; // don't load unless required        
-        String url = cache.getContentUrl(cachedContentFile);
-        if (url == null)
+        if (targetReductionBytes > 0 && sizeFilesDeleted < targetReductionBytes)
         {
-            // Not in the cache, check the properties file
-            props = new CacheFileProps(cachedContentFile);
-            props.load();
-            url = props.getContentUrl();
-        }   
+            // Aggressive clean mode, delete file straight away.
+            deleted = deleteFilesNow(cachedContentFile);
+        }
+        else
+        {
+            if (oldEnoughForCleanup(cachedContentFile))
+            {
+                if (log.isDebugEnabled())
+                {
+                    log.debug("File is older than " + minFileAgeMillis + 
+                                "ms - considering for cleanup: " + cachedContentFile);
+                }
+                props = new CacheFileProps(cachedContentFile);        
+                String url = cache.getContentUrl(cachedContentFile);
+                if (url == null)
+                {
+                    // Not in the cache, check the properties file 
+                    props.load();
+                    url = props.getContentUrl();
+                }   
+                
+                if (url == null || !cache.contains(url))
+                {
+                    // If the url is null, it might still be in the cache, but we were unable to determine it
+                    // from the reverse lookup or the properties file. Delete the file as it is most likely orphaned.
+                    // If for some reason it is still in the cache, cache.getReader(url) must re-cache it.
+                    deleted = markOrDelete(cachedContentFile, props);
+                }
+            }
+            else
+            {
+                if (log.isDebugEnabled())
+                {
+                    log.debug("File too young for cleanup - ignoring " + cachedContentFile);
+                }
+            }
+        }
         
-        if (url != null && !cache.contains(url))
+        if (!deleted)
         {
             if (props == null)
             {
                 props = new CacheFileProps(cachedContentFile);
-                props.load();
             }
-            markOrDelete(cachedContentFile, props);
+            long size = cachedContentFile.length() + props.fileSize();
+            newDiskUsage += size;
         }
-        else if (url == null)
-        {
-            // It might still be in the cache, but we were unable to determine it from the reverse lookup
-            // or the properties file. Delete the file as it is most likely orphaned. If for some reason it is
-            // still in the cache, cache.getReader(url) must re-cache it.
-            markOrDelete(cachedContentFile, props);
-        }    
     }
 
     
+
+    /**
+     * Is the file old enough to be considered for cleanup/deletion? The file must be older than minFileAgeMillis
+     * to be considered for deletion - the state of the cache and the file's associated properties file will not
+     * be examined unless the file is old enough.
+     *  
+     * @return true if the file is older than minFileAgeMillis, false otherwise.
+     */
+    private boolean oldEnoughForCleanup(File file)
+    {
+        if (minFileAgeMillis == 0)
+        {
+            return true;
+        }
+        else
+        {
+            long now = System.currentTimeMillis();
+            return (file.lastModified() < (now - minFileAgeMillis));
+        }
+    }
+
 
     /**
      * Marks a file for deletion by a future run of the CachedContentCleaner. Each time a file is observed
@@ -99,8 +248,9 @@ public class CachedContentCleaner implements FileHandler
      * 
      * @param file
      * @param props
+     * @return true if the content file was deleted, false otherwise.
      */
-    private void markOrDelete(File file, CacheFileProps props)
+    private boolean markOrDelete(File file, CacheFileProps props)
     {
         Integer deleteWatchCount = props.getDeleteWatchCount();
 
@@ -108,16 +258,30 @@ public class CachedContentCleaner implements FileHandler
         if (deleteWatchCount < 0)
             deleteWatchCount = 0;
         
+        boolean deleted = false;
+        
         if (deleteWatchCount < maxDeleteWatchCount)
         {
             deleteWatchCount++;
+            
+            if (log.isDebugEnabled())
+            {
+                log.debug("Marking file for deletion, deleteWatchCount=" + deleteWatchCount + ", file: "+ file);
+            }
             props.setDeleteWatchCount(deleteWatchCount);
             props.store();
+            numFilesMarked++;
         }
         else
         {
-            deleteFilesNow(file);
+            if (log.isDebugEnabled())
+            {
+                log.debug("Deleting cache file " + file);
+            }
+            deleted = deleteFilesNow(file);
         }
+        
+        return deleted;
     }
 
     /**
@@ -125,13 +289,22 @@ public class CachedContentCleaner implements FileHandler
      * original content URL and deletion marker information.
      *  
      * @param cacheFile Location of cached content file.
+     * @return true if the content file was deleted, false otherwise.
      */
-    private void deleteFilesNow(File cacheFile)
+    private boolean deleteFilesNow(File cacheFile)
     {
         CacheFileProps props = new CacheFileProps(cacheFile);
         props.delete();
-        cacheFile.delete();
-        Deleter.deleteEmptyParents(cacheFile, cache.getCacheRoot());
+        long fileSize = cacheFile.length();
+        boolean deleted = cacheFile.delete();
+        if (deleted)
+        {
+            numFilesDeleted++;
+            sizeFilesDeleted += fileSize;
+            Deleter.deleteEmptyParents(cacheFile, cache.getCacheRoot());
+        }
+        
+        return deleted;
     }
 
     
@@ -142,6 +315,24 @@ public class CachedContentCleaner implements FileHandler
         this.cache = cache;
     }
 
+    
+    /**
+     * Sets the minimum age of a cache file before it will be considered for deletion.
+     * @see #oldEnoughForCleanup(File)
+     * @param minFileAgeMillis
+     */
+    public void setMinFileAgeMillis(long minFileAgeMillis)
+    {
+        this.minFileAgeMillis = minFileAgeMillis;
+    }
+
+
+    /**
+     * Sets the maxDeleteWatchCount value.
+     * 
+     * @see #markOrDelete(File, CacheFileProps)
+     * @param maxDeleteWatchCount
+     */
     public void setMaxDeleteWatchCount(Integer maxDeleteWatchCount)
     {
         if (maxDeleteWatchCount < 0)
@@ -149,5 +340,87 @@ public class CachedContentCleaner implements FileHandler
             throw new IllegalArgumentException("maxDeleteWatchCount cannot be negative [value=" + maxDeleteWatchCount + "]");
         }
         this.maxDeleteWatchCount = maxDeleteWatchCount;
+    }
+
+
+    /**
+     * @param usageTracker the usageTracker to set
+     */
+    public void setUsageTracker(UsageTracker usageTracker)
+    {
+        this.usageTracker = usageTracker;
+    }
+
+    public boolean isRunning()
+    {
+        lock.readLock().lock();
+        try
+        {
+            return running;
+        }
+        finally
+        {
+            lock.readLock().unlock();
+        }
+    }
+    
+    public long getNumFilesSeen()
+    {
+        return this.numFilesSeen;
+    }
+    
+    public long getNumFilesDeleted()
+    {
+        return this.numFilesDeleted;
+    }
+
+    public long getSizeFilesDeleted()
+    {
+        return this.sizeFilesDeleted;
+    }
+        
+    public double getSizeFilesDeletedMB()
+    {
+        return (double) getSizeFilesDeleted() / FileUtils.ONE_MB;
+    }
+
+    public long getNumFilesMarked()
+    {
+        return numFilesMarked;
+    }
+    
+    public Date getTimeStarted()
+    {
+        return this.timeStarted;
+    }
+
+    public Date getTimeFinished()
+    {
+        return this.timeFinished;
+    }
+    
+    public long getDurationSeconds()
+    {
+        return getDurationMillis() / 1000;
+    }
+    
+    public long getDurationMillis()
+    {
+        return timeFinished.getTime() - timeStarted.getTime();
+    }
+
+    @Override
+    public void setApplicationEventPublisher(ApplicationEventPublisher eventPublisher)
+    {
+        this.eventPublisher = eventPublisher;
+    }
+
+    /**
+     * Returns the cacheRoot that this cleaner is responsible for.
+     * @return File
+     */
+    public File getCacheRoot()
+    {
+        return cache.getCacheRoot();
     }
 }
